services/order-service/requirements.txt
flask
mysql-connector-python
flask-cors
requests
fpdf 